requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
psycopg2-binary
flask-cors
